# Conflicts:
#	includes/rt_opencl_params_defines.h
#	sources/init/rt_init.c
#	sources/kernel/object_intersections/intersection_utils.cl
#	sources/kernel/object_intersections/more_difficult_object/unlimited_object.cl
#	sources/kernel/object_intersections/simple_object.cl
#	sources/kernel/prototypes.cl
#	sources/rt_parsing/hardcoded_scene.c

+++ sources/kernel/color_utils.cl
@@ -39,7 +39,7 @@ int			get_int_color(float3 color)
 /*
 int         get_texture_color(
 		int2 pos,
-		__constant float *texture_list,
+		__constant int *texture_list,
 		__constant t_texture_info *texture_info
 		)
 {
